Rails 7.1
We're disabling `action_controller.raise_on_missing_callback_actions`
because sometimes we include `before_action :something, only: actions`
in concerns, and we include these concerns in controllers that don't
have all these actions.

Note that Rails 7.1 logs to STDOUT by default [1]; we're re-adding the
condition `if ENV["RAILS_LOG_TO_STDOUT"].present?` because we're still
using files and we're rotating the logs to avoid running out of space.

Also note that, since now the `shared_extensions` schema is added to the
`schema.rb` file, we need to make sure we move extensions to that schema
_after_ the `db:schema:load` task has finished.

[1] Pull request 47138 in https://github.com/rails/rails